[ENG-1630] Share add registration provider info (#9348)
## Purpose

This ticket should ensure the access token is passed from the node provider object to SHARE.

## Changes

So I had to renovate some of this code, because there was multiple bugs in the tests and because the code itself was old and doesn't fit our current conventions. So I 

0) Moved `access_token` field to `AbstractProvider` instead of previous weird situation
1) Moved node license tests to `osf_tests/test_node_license.py`
2) Move node SHARE tests to `osf_tests/test_node_share.py`
3) Fixed and DRY-ed up most of the mocking for `test_node_share.py`

## QA Notes

Dev test only.

## Documentation

tech task, no new docs.

## Side Effects

Detours:

## Ticket

https://openscience.atlassian.net/browse/ENG-1630
